% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_editor_context.R
\name{convert_selection}
\alias{convert_selection}
\title{Convert and load the hightlighted assignment to the global environment}
\usage{
convert_selection()
}
\description{
After highlighting the assignment in the source editor, go to 
the console and run this function. The selected code will be run 
and if it is reactive, it will be loaded as a funciton.
}
